feat: streamline unneeded query transaction usage
