+++ src/metakb/cli.py
@@ -15,11 +15,13 @@ import asyncclick as click
 import boto3
 from boto3.exceptions import ResourceLoadException
 from botocore.config import Config
+from neo4j import Driver
 
 from metakb import APP_ROOT, DATE_FMT
-from metakb.database import Graph
+from metakb.database import get_driver
 from metakb.harvesters.civic import CivicHarvester
 from metakb.harvesters.moa import MoaHarvester
+from metakb.load_data import load_from_json
 from metakb.log_handle import configure_logs
 from metakb.normalizers import (
     NORMALIZER_AWS_ENV_VARS,
@@ -326,8 +328,8 @@ async def transform_file(
     )
 
 
-def _get_graph(db_url: str, db_creds: str | None) -> Generator[Graph, None, None]:
-    """Acquire Neo4j graph.
+def _get_driver(db_url: str, db_creds: str | None) -> Generator[Driver, None, None]:
+    """Acquire Neo4j graph driver.
 
     :param db_url: URL endpoint for the application Neo4j database.
     :param db_creds: DB username and password, separated by a colon, e.g.
@@ -342,11 +344,11 @@ def _get_graph(db_url: str, db_creds: str | None) -> Generator[Graph, None, None
             credentials = (split_creds[0], split_creds[1])
         except IndexError:
             _help_msg(
-                f"Argument to --db_credentialss appears invalid. Got '{db_creds}'. Should follow pattern 'username:password'."
+                f"Argument to --db_credentials appears invalid. Got '{db_creds}'. Should follow pattern 'username:password'."
             )
-    graph = Graph(uri=db_url, credentials=credentials)
-    yield graph
-    graph.close()
+    driver = get_driver(uri=db_url, credentials=credentials)
+    yield driver
+    driver.close()
 
 
 @cli.command()
@@ -369,8 +371,8 @@ def clear_graph(db_url: str, db_credentials: str | None) -> None:
     :param db_credentials: DB username and password, separated by a colon, e.g.
         ``"username:password"``.
     """  # noqa: D301
-    graph = next(_get_graph(db_url, db_credentials))
-    graph.clear()
+    driver = next(_get_driver(db_url, db_credentials))
+    clear_graph(driver)
 
 
 @cli.command()
@@ -428,11 +430,11 @@ def load_cdm(
     start = timer()
     _echo_info("Loading Neo4j database...")
 
-    graph = next(_get_graph(db_url, db_credentials))
+    driver = next(_get_driver(db_url, db_credentials))
 
     if cdm_files:
         for file in cdm_files:
-            graph.load_from_json(file)
+            load_from_json(file, driver)
     else:
         version = _retrieve_s3_cdms() if from_s3 else "*"
 
@@ -446,7 +448,7 @@ def load_cdm(
                 msg = f"No valid transform file found matching pattern: {pattern}"
                 raise FileNotFoundError(msg) from e
 
-            graph.load_from_json(path)
+            load_from_json(path, driver)
 
     end = timer()
     _echo_info(f"Successfully loaded neo4j database in {(end - start):.5f} s")
@@ -515,7 +517,7 @@ async def update(
     start = timer()
     _echo_info("Loading Neo4j database...")
 
-    graph = next(_get_graph(db_url, db_credentials))
+    driver = next(_get_driver(db_url, db_credentials))
 
     if not sources:
         sources = tuple(SourceName)
@@ -529,9 +531,9 @@ async def update(
             msg = f"No valid transform file found matching pattern: {pattern}"
             raise FileNotFoundError(msg) from e
 
-        graph.load_from_json(path)
+        load_from_json(path, driver)
 
-    graph.close()
+    driver.close()
     end = timer()
     _echo_info(f"Successfully loaded neo4j database in {(end - start):.5f} s")
